+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Release Notes
+## [v1.1.3](https://github.com/404Exist/datatable-cruds/compare/v1.1.2...v1.1.3) - 2022-12-17
+
+### Added
+- `php artisan datatablecruds:for {Model}` command has been added
+### Fixed
+
+- IDE autocompelete methods names issue fixed
+- fixQueryWheresColumnsNames in ModelDataTable fixed
+- Order & search by relations of relation fixed
+### Removed
+
+- `editor()` input has been removed because it's script is too large and you can set your custom one if needed.
 ## [v1.1.2](https://github.com/404Exist/datatable-cruds/compare/v1.1.1...v1.1.2) - 2022-12-16
 
 ### Updated
+
 - `setModel()` updated to `for()` and it now accepts `Model::query()` or `Model::class`
 ## [v1.1.1](https://github.com/404Exist/datatable-cruds/compare/v1.1.0...v1.1.1) - 2022-12-16
 
@@ -11,11 +24,12 @@
 - `columns()` & `inputs()` methods have been added.
 - `showPagination()` & `hidePaginationIfContainOnePage()` methods have been added.
 - `datatableCruds()` helper method have been added, it returns an instance of DatatableCruds class.
-- Access {current_page} and {per_page} in custom exports html.
+- Access |current_page| and |per_page| in custom exports html.
 - Ability to set exported filename.
 - Ability to send model query to `dataTableOf()` helper function.
 - Ability to set nested text data with `setText('delete.title', 'Delete')`.
 ### Fixed
+
 - `setBladeExtends()` & `setBladeSection()` methods have been fixed.
 ## [v1.1.0](https://github.com/404Exist/datatable-cruds/compare/v1.0.9...v1.1.0) - 2022-12-15
 
@@ -49,6 +63,7 @@
 
 
 ### Removed
+
 - `fillColumns()` method have been removed.
 - `fillInputs()` method have been removed.
 - `sort()` method have been removed.
@@ -63,8 +78,10 @@
 ## [v1.0.6](https://github.com/404Exist/datatable-cruds/compare/v1.0.5...v1.0.6) - 2022-04-10
 
 ### Added
+
 - `setBladeExtends()` and `setBladeSection()` methods have been added.
 ### Updated
+
 - `render()` method now accepts one parameter only which is an array of variables that can be accessed in the layout blade file.
 ## [v1.0.5](https://github.com/404Exist/datatable-cruds/compare/v1.0.3...v1.0.5) - 2022-04-9